+++ assets/asset-netlify/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/asset-netlify
 
+## 1.0.0-beta.16
+
+### Patch Changes
+
+- df
+
 ## 1.0.0-beta.15
 
 ### Patch Changes