+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pipelab/shared
 
+## 1.0.0-beta.16
+
+### Patch Changes
+
+- df
+- Updated dependencies
+  - @pipelab/constants@1.0.0-beta.18
+  - @pipelab/migration@1.0.0-beta.16
+
 ## 1.0.0-beta.15
 
 ### Patch Changes