+++ plugins/plugin-system/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pipelab/plugin-system
 
+## 1.0.0-beta.21
+
+### Patch Changes
+
+- df
+- Updated dependencies
+  - @pipelab/plugin-core@1.0.0-beta.21
+
 ## 1.0.0-beta.20
 
 ### Patch Changes